update delete button

+++ src/pages/ProductDetails.tsx
@@ -9,7 +9,7 @@
 // export default function ProductDetails() {
 //     const {id} =useParams()
 //     const {data:product,isLoading,error} = useGetProductDetailsQuery(id)
-  
+
 //   return (
 //     <>
 //     <Navbar />
@@ -34,66 +34,74 @@
 //   )
 // }
 
-import React from 'react';
-import ProductReview from './ProductReview';
-import { useDeleteProductMutation, useGetProductDetailsQuery } from '../redux/feature/product/apiSlice';
-import { Link, useParams } from 'react-router-dom';
-import Navbar from '../layouts/Navbar';
-import Footer from '../layouts/Footer';
-import { useAppDispatch } from '../redux/hook';
-
-
+import React, { useState } from "react";
+import ProductReview from "./ProductReview";
+import {
+  useDeleteProductMutation,
+  useGetProductDetailsQuery,
+} from "../redux/feature/product/apiSlice";
+import { Link, useParams } from "react-router-dom";
+import Navbar from "../layouts/Navbar";
+import Footer from "../layouts/Footer";
+import { useAppDispatch } from "../redux/hook";
 
 export default function ProductDetails() {
+  const [bookDeleted, setBookDeleted] = useState(false);
   const { id } = useParams<{ id: string }>();
   const { data: product, isLoading, error } = useGetProductDetailsQuery(id);
-  const [deleteProduct, { isLoading:isload, isError, isSuccess }] =
+  const [deleteProduct, { isLoading: isload, isError, isSuccess }] =
     useDeleteProductMutation();
-    console.log(isload)
-    console.log(isError)
-    console.log(isSuccess)
-  //const dispatch = useAppDispatch();
+  console.log(isload);
+  console.log(isError);
+  console.log(isSuccess);
 
   const handleDeleteProduct = () => {
-    if (id) {
-      deleteProduct(id);
+    const confirmed = window.confirm(
+      "Are you sure you want to delete this book?"
+    );
+    if (confirmed) {
+      if (id) {
+        deleteProduct(id);
+      }
+      setBookDeleted(true);
     }
-    console.log(id)
-    
+
+    console.log(id);
   };
 
   return (
     <>
-      
       <div className="flex max-w-7xl mx-auto items-center border-b border-gray-300">
         <div className="w-[50%]">
-          <p className="text-3xl font-semibold">{product?.title}</p>
+          {/* <p className="text-3xl font-semibold">{product?.title}</p> */}
         </div>
         <div className="w-[50%] space-y-3">
-          <h1 className="text-3xl font-semibold">Author Name: {product?.author}</h1>
-          <p className="text-xl">Genre: {product?.genre}</p>
-          <ul className="space-y-1 text-lg">
+          <p className="text-2xl font-semibold">Booke Name:{product?.title}</p>
+          <h1 className="text-2xl font-semibold">
+            Author Name: {product?.author}
+          </h1>
+          <p className="text-2xl">Genre: {product?.genre}</p>
+          <ul className="space-y-1 text-2xl">
             Published Date: {product?.published_date}
           </ul>
-          <div className="card-actions justify-end">
-            <button className="btn" onClick={handleDeleteProduct} >
+          <div className="card-actions justify-center">
+            <button className="btn">
+              <Link to={`/updateproduct/${id}`}>Edit Book</Link>
+            </button>
+            {/* <button className="btn" onClick={handleDeleteProduct} >
              Delete Book
-            </button>
-          </div>
-          <div className="card-actions justify-first">
-            <button className="btn"  >
-             <Link to={`/updateproduct/${id}`}>
-             Edit Book
-             </Link>
-            </button>
+            </button> */}
+            {bookDeleted ? (
+              <p>Book has been deleted.</p>
+            ) : (
+              <button className="btn" onClick={handleDeleteProduct}>
+                Delete Book
+              </button>
+            )}
           </div>
         </div>
       </div>
-      <ProductReview id={id!}/>
-      
+      <ProductReview id={id!} />
     </>
   );
 }
-
-
-
